Django/day_2/pyproject.toml
[tool.poetry]
name = "oz-day-2"
version = "0.1.0"
description = ""
authors = ["gunung-kim"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
ipython = "^9.11.0"
django-extensions = "^4.1"
django-summernote = "^0.8.20.0"
pillow = "^12.1.1"
django-cleanup = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
